refactor: Change name of AppUserDetails and add annotations in AppUserDetailsService
