+++ src/main/java/com/app/housing_association/security/login/service/AppUserDetailsService.java
@@ -1,22 +1,28 @@
-package com.app.housing_association.security.login;
+package com.app.housing_association.security.login.service;
 
-import com.app.housing_association.security.login.model.UserEntityDetails;
+import com.app.housing_association.security.login.model.AppUserDetails;
 import com.app.housing_association.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import static java.lang.String.format;
 
 @RequiredArgsConstructor
+@Service
 public class AppUserDetailsService implements UserDetailsService {
 
     private final UserRepository userRepository;
 
     @Override
+    @Transactional
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
         return userRepository
                 .findByUsernameIgnoreCase(username)
-                .map(UserEntityDetails::new)
-                .orElseThrow(() -> new UsernameNotFoundException(username));
+                .map(AppUserDetails::new)
+                .orElseThrow(() -> new UsernameNotFoundException(format("User %s, not found", username)));
     }
 }
